Reading files one at a time to database
