+++ col_count.csv
@@ -1,55 +1,28 @@
 col_name,count
-status,147
-family_type,125
-adult,147
-infant,147
-preshooler,85
-schoolager,147
-teenager,147
-state,125
-year,127
-all_families_table_#,125
-county,124
-housing_costs,124
-child_care_costs,124
-food_costs,124
-transportation_costs,124
-health_care_costs,123
-miscellaneous_costs,124
-taxes,147
-earned_income_tax_credit,146
-child_care_tax_credit,146
-child_tax_credit,146
-hourly_self-sufficiency_wage,147
-monthly_self-sufficiency_wage,147
-annual_self-sufficiency_wage,147
-emergency_savings,147
-id,22
-release_year,21
-state_name,22
-state1,22
-county1,22
-famcode,22
-preschooler,62
-housing,23
-child_care,23
-food,23
-health_care,23
-transportation,23
-miscellaneous,23
-broadband_&_cell_phone,9
-other_necessities,9
-health_care__costs,1
-premium,1
-out-of-pocket,1
-federal_income_taxes,1
-payroll_taxes,1
-state_sales_taxes,1
-state_income_taxes,1
-federal_child_tax_credit,1
-federal_&_oregon_eitc,1
-federal_cdctc,1
-oregon_wfhdc,1
-total_annual_resources,1
-town,1
-working_family_household_and_dependent_care_credit,1
+status,2
+family_type,2
+adult,2
+infant,2
+preschooler,2
+schoolager,2
+teenager,2
+state,2
+year,2
+all_families_table_#,1
+place,2
+housing,2
+child_care,2
+food,2
+transportation,2
+health_care,2
+miscellaneous,2
+taxes,2
+earned_income_tax_credit,2
+child_care_tax_credit,2
+child_tax_credit,2
+hourly_self_sufficiency_wage,2
+monthly_self_sufficiency_wage,2
+annual_self_sufficiency_wage,2
+emergency_savings,2
+id,1
+state_name,1
